requirements.txt
discord
cohere
python-dotenv
flask
firebase_admin